updates, working cooldown test, working action reg write to fire external DGGs (internal not working, see main_DGG_debug), WIP on reading out LRS 4300B. Currently reading 0's

+++ UserTools/camacinc/CamacCrate/CamacCrate.h
@@ -43,12 +43,12 @@ class CamacCrate
 		int GetSlot(int ID);
 		void ListSlot();
 		//PMT Test Stand Functions
-		short DelayGateGen(short chan, short trig, short out, short delay, short gate, short invert, short latch);
+		short DelayGateGen(short chan, short trig, short out, int delay, int gate, short invert, short latch);
 		short RegWrite(short A, long Data);
 		short RegRead(short A, long &Data);
 		short CC_NIM_Out(short A, short F, short inv, short latch);
-		int ActionRegWrite(long value);
-		int ActionRegRead(long &Data);
+		short ActionRegWrite(long value);
+		short ActionRegRead(long &Data);
 
 		int Z();		//Z
 		int C();		//C
